add set grenade pos and remove extra set_velocity
Moving the object away from the block it's hitting nearly eliminates the erratic behavior when landing. before when it lands it usually flies off in a seemingly random direction. I think this is related to https://github.com/minetest/minetest/issues/5330

Also removed the extra set_velocity and get_velocity and just changed the vel variable. It works out the same.